+++ services/bing.js
@@ -15,7 +15,7 @@ const bingImageSearch = search => {
     params: { q: `${encodeURIComponent(search)}` },
     headers: {
       "x-rapidapi-key": process.env.API_KEY,
-      "x-rapidapi-host": process.env.API_HOST
+      "x-rapidapi-host": "bing-image-search1.p.rapidapi.com"
     }
   };
 
@@ -29,19 +29,11 @@ const bingImageSearch = search => {
       const imageALT = imageArray.map(value => value.name);
       const imageSRC = imageArray.map(value => value.thumbnailUrl);
       const imageHREF = imageArray.map(value => value.hostPageUrl);
-      // const imgStringArray = [];
-      // for (index = 0; index < imageArray.length; index++) {
-      //   imgStringArray.push(
-      //     `<img alt="${imageALT[index]}" src="${imageSRC[index]}" href="${imageHREF[index]}">`
-      //   );
-      //   // document.body.appendChild(imgStringArray);
-      // }
-      // console.log(imgStringArray);
       // - construct image objects from 3 arrays;
       const imageObj = imageALT.map((item, index) => ({
         alt: item,
         src: imageSRC[index] || "",
-        link: imageHREF[index] || ""
+        href: imageHREF[index] || ""
       }));
       const shuffledImageObj = shuffle(imageObj);
       // console.log(shuffledImageObj);
@@ -52,8 +44,4 @@ const bingImageSearch = search => {
     });
 };
 // - Exports;
-// bingImageSearch("90s");
 module.exports = bingImageSearch;
-
-// var myImage = new Image(100, 200);
-// myImage.src = 'picture.jpg';
